nix-profile-daemon.fish: Check $NIX_STATE_HOME for ~/.nix-profile
There seems to be no reason to have different logic when checking for
the `~/.nix-profile` location when nix is a daemon vs when it is run as
a user process.  If setting `$NIX_STATE_HOME` is useful in once case, it
is probably useful in the other case as well?

Plus, differences between `nix-profile.fish` and
`nix-profile-daemon.fish` complicate maintenance, so it probably makes
sense to keep them as similar as possible.