doc: Switch installation examples to use temporary file
The installation script checks if the stdin is a terminal, and if it is
not, it switches to a non-interactive mode.  Considering these examples
expect an interactive installation, it is better to use the same code as
in `doc/manual/source/installation/installing-binary.md`, where the
installation script is downloaded into a temporary file.